requirements.txt
# OpenClaw backend — Oracle Cloud ARM VM
# Install with: pip install -r requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
uvloop>=0.19.0

# Telegram
python-telegram-bot>=21.0

# HTTP client (Ollama + HF calls)
httpx>=0.27.0

# Async SQLite job queue
aiosqlite>=0.20.0

# Scheduler
apscheduler>=3.10.4

# Config
python-dotenv>=1.0.0
